+++ dependify/decorators/__init__.py
@@ -1,0 +1,11 @@
+__all__ = [
+    "wired",
+    "injectable",
+    "inject",
+    "injected",
+]
+
+from dependify.decorators._inject import inject
+from dependify.decorators._injectable import injectable
+from dependify.decorators._injected import injected
+from dependify.decorators._wired import wired
