+++ dependify/decorators/_inject/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "inject"
+]
+
+from dependify.decorators._inject._inject import inject
