+++ dependify/decorators/_injected/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "injected"
+]
+
+from dependify.decorators._injected._injected import injected
