+++ dependify/decorators/_injected/_injected.py
@@ -1,0 +1,80 @@
+from functools import partial
+from inspect import Parameter, Signature
+from typing import Callable, Optional, TypeVar, Union
+
+from dependify.context import registry
+from dependify.decorators._injected._get_annotations import get_annotations
+from dependify.dependency_registry import DependencyRegistry
+
+class_type = TypeVar("class_type", bound=type)
+
+
+def injected(
+    class_: Optional[class_type] = None,
+    *,
+    validate: bool = True,
+    registry_: DependencyRegistry = registry,
+) -> Union[class_type, Callable[[class_type], class_type]]:
+    """
+    Decorator to create default constructor of a class it none present
+    """
+    if class_ is None:
+        return partial(injected, registry=registry_, validate=validate)
+    if "__init__" in class_.__dict__:
+        return class_
+    class_annotations = get_annotations(class_)
+    annotations = tuple(class_annotations.items())
+
+    def __init__(self, *args, **kwargs):
+        annotations_iter = iter(annotations)
+        for arg, (field_name, field_type) in zip(args, annotations_iter):
+            if (
+                validate
+                and isinstance(field_type, type)
+                and not isinstance(arg, field_type)
+            ):
+                raise TypeError(
+                    f"Expected {field_type} for {field_name}, got {type(arg)}"
+                )
+            setattr(self, field_name, arg)
+        missing_args = set(field_name for field_name, _ in annotations_iter)
+        kwargs_copy = kwargs.copy()
+        for field_name, value in tuple(kwargs_copy.items()):
+            if field_name not in class_annotations:
+                raise TypeError(
+                    f"Keyword argument: {field_name} not found in class {class_.__name__}"
+                )
+            if field_name not in missing_args:
+                raise TypeError(
+                    f"Keyword argument: {field_name} already provided as a positional argument"
+                )
+            field_type = class_annotations[field_name]
+            if (
+                validate
+                and isinstance(field_type, type)
+                and not isinstance(value, field_type)
+            ):
+                raise TypeError(
+                    f"Expected {field_type} for {field_name}, got {type(value)}"
+                )
+            setattr(self, field_name, value)
+            missing_args.remove(field_name)
+        for field_name in tuple(missing_args):
+            if hasattr(class_, field_name):
+                setattr(self, field_name, getattr(class_, field_name))
+                missing_args.remove(field_name)
+        if missing_args:
+            raise TypeError(f"Missing arguments: {', '.join(missing_args)}")
+
+    __init__.__annotations__ = class_annotations.copy()
+    __init__.__signature__ = Signature(
+        (Parameter("self", Parameter.POSITIONAL_OR_KEYWORD),)
+        + tuple(
+            Parameter(
+                name, Parameter.POSITIONAL_OR_KEYWORD, annotation=annotation
+            )
+            for name, annotation in __init__.__annotations__.items()
+        )
+    )
+    class_.__init__ = inject(__init__, registry_=registry_)
+    return class_
