+++ dependify/dependencies/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "Dependency",
+]
+
+from dependify import Dependency
