+++ test/unit/test_inject.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from unittest.mock import patch
 
-from src.dependify import inject
+from dependify import inject
 
 
 class TestInject(TestCase):
@@ -29,7 +29,7 @@ class TestInject(TestCase):
         """
         Test if the `inject` decorator can be used with a custom registry.
         """
-        from src.dependify import DependencyRegistry
+        from dependify import DependencyRegistry
 
         class A:
             pass
